Refactor hardhat helpers
